+++ lib.esm/index.mjs
@@ -7082,6 +7082,39 @@ class ZGServingUserBrokerBase {
             throw error;
         }
     }
+    async getQuote(providerAddress) {
+        try {
+            const service = await this.getService(providerAddress);
+            const url = service.url;
+            const endpoint = `${url}/v1/quote`;
+            const quoteString = await this.fetchText(endpoint, {
+                method: 'GET',
+            });
+            const ret = JSON.parse(quoteString, (_, value) => {
+                if (typeof value === 'string' && /^\d+$/.test(value)) {
+                    return BigInt(value);
+                }
+                return value;
+            });
+            return ret;
+        }
+        catch (error) {
+            throw error;
+        }
+    }
+    async fetchText(endpoint, options) {
+        try {
+            const response = await fetch(endpoint, options);
+            if (!response.ok) {
+                throw new Error(`HTTP error! Status: ${response.status}`);
+            }
+            const buffer = await response.arrayBuffer();
+            return Buffer.from(buffer).toString('utf-8');
+        }
+        catch (error) {
+            throw error;
+        }
+    }
     async getExtractor(providerAddress, useCache = true) {
         try {
             const svc = await this.getService(providerAddress, useCache);
@@ -8475,158 +8508,6 @@ class Automata {
 }
 
 /**
- * RequestProcessor is a subclass of ZGServingUserBroker.
- * It needs to be initialized with createZGServingUserBroker
- * before use.
- */
-class RequestProcessor extends ZGServingUserBrokerBase {
-    automata;
-    constructor(contract, metadata, cache, ledger) {
-        super(contract, ledger, metadata, cache);
-        this.automata = new Automata();
-    }
-    async getServiceMetadata(providerAddress) {
-        const service = await this.getService(providerAddress);
-        return {
-            endpoint: `${service.url}/v1/proxy`,
-            model: service.model,
-        };
-    }
-    /*
-     * 1. To Ensure No Insufficient Balance Occurs.
-     *
-     * The provider settles accounts regularly. In addition, we will add a rule to the provider's settlement logic:
-     * if the actual balance of the customer's account is less than 5000, settlement will be triggered immediately.
-     * The actual balance is defined as the customer's inference account balance minus any unsettled amounts.
-     *
-     * This way, if the customer checks their account and sees a balance greater than 5000, even if the provider settles
-     * immediately, the deduction will leave about 5000, ensuring that no insufficient balance situation occurs.
-     *
-     * 2. To Avoid Frequent Transfers
-     *
-     * On the customer's side, if the balance falls below 5000, it should be topped up to 10000. This is to avoid frequent
-     * transfers.
-     *
-     * 3. To Avoid Having to Check the Balance on Every Customer Request
-     *
-     * Record expenditures in processResponse and maintain a total consumption amount. Every time the total expenditure
-     * reaches 1000, recheck the balance and perform a transfer if necessary.
-     *
-     * ps: The units for 5000 and 1000 can be (service.inputPricePerToken + service.outputPricePerToken).
-     */
-    async getRequestHeaders(providerAddress, content, vllmProxy) {
-        try {
-            await this.topUpAccountIfNeeded(providerAddress, content);
-            if (!vllmProxy) {
-                vllmProxy = false;
-            }
-            return await this.getHeader(providerAddress, content, BigInt(0), vllmProxy);
-        }
-        catch (error) {
-            throw error;
-        }
-    }
-    async acknowledgeProviderSigner(providerAddress, gasPrice) {
-        try {
-            try {
-                await this.contract.getAccount(providerAddress);
-            }
-            catch (error) {
-                if (!error.message.includes('AccountNotExists')) {
-                    throw error;
-                }
-                else {
-                    await this.ledger.transferFund(providerAddress, 'inference', BigInt(0), gasPrice);
-                }
-            }
-            let { quote, provider_signer, key } = await this.getQuote(providerAddress);
-            if (!quote || !provider_signer) {
-                throw new Error('Invalid quote');
-            }
-            if (!quote.startsWith('0x')) {
-                quote = '0x' + quote;
-            }
-            const rpc = process.env.RPC_ENDPOINT;
-            // bypass quote verification if testing on localhost
-            if (!rpc || !/localhost|127\.0\.0\.1/.test(rpc)) {
-                const isVerified = await this.automata.verifyQuote(quote);
-                console.log('Quote verification:', isVerified);
-                if (!isVerified) {
-                    throw new Error('Quote verification failed');
-                }
-            }
-            const account = await this.contract.getAccount(providerAddress);
-            if (account.providerPubKey[0] === key[0] &&
-                account.providerPubKey[1] === key[1]) {
-                console.log('Provider signer already acknowledged');
-                return;
-            }
-            await this.contract.acknowledgeProviderSigner(providerAddress, key);
-            const userAddress = this.contract.getUserAddress();
-            const cacheKey = `${userAddress}_${providerAddress}_ack`;
-            await this.cache.setItem(cacheKey, key, 1 * 60 * 1000, CacheValueTypeEnum.Other);
-        }
-        catch (error) {
-            throw error;
-        }
-    }
-    async getQuote(providerAddress) {
-        try {
-            const service = await this.getService(providerAddress);
-            const url = service.url;
-            const endpoint = `${url}/v1/quote`;
-            const quoteString = await this.fetchText(endpoint, {
-                method: 'GET',
-            });
-            const ret = JSON.parse(quoteString, (_, value) => {
-                if (typeof value === 'string' && /^\d+$/.test(value)) {
-                    return BigInt(value);
-                }
-                return value;
-            });
-            return ret;
-        }
-        catch (error) {
-            throw error;
-        }
-    }
-    async fetchText(endpoint, options) {
-        try {
-            const response = await fetch(endpoint, options);
-            if (!response.ok) {
-                throw new Error(`HTTP error! Status: ${response.status}`);
-            }
-            const buffer = await response.arrayBuffer();
-            return Buffer.from(buffer).toString('utf-8');
-        }
-        catch (error) {
-            throw error;
-        }
-    }
-}
-
-var VerifiabilityEnum;
-(function (VerifiabilityEnum) {
-    VerifiabilityEnum["OpML"] = "OpML";
-    VerifiabilityEnum["TeeML"] = "TeeML";
-    VerifiabilityEnum["ZKML"] = "ZKML";
-})(VerifiabilityEnum || (VerifiabilityEnum = {}));
-let ModelProcessor$1 = class ModelProcessor extends ZGServingUserBrokerBase {
-    async listService() {
-        try {
-            const services = await this.contract.listService();
-            return services;
-        }
-        catch (error) {
-            throw error;
-        }
-    }
-};
-function isVerifiability(value) {
-    return Object.values(VerifiabilityEnum).includes(value);
-}
-
-/**
  * The Verifier class contains methods for verifying service reliability.
  */
 class Verifier extends ZGServingUserBrokerBase {
@@ -8651,7 +8532,7 @@ class Verifier extends ZGServingUserBrokerBase {
      * @returns The first return value indicates whether the RA is valid,
      * and the second return value indicates the signing address of the RA.
      */
-    async getSigningAddress(providerAddress, verifyRA = false) {
+    async getSigningAddress(providerAddress, verifyRA = false, vllmProxy = false) {
         const key = `${this.contract.getUserAddress()}_${providerAddress}`;
         let signingKey = await this.metadata.getSigningKey(key);
         if (!verifyRA && signingKey) {
@@ -8663,9 +8544,24 @@ class Verifier extends ZGServingUserBrokerBase {
         try {
             const extractor = await this.getExtractor(providerAddress, false);
             const svc = await extractor.getSvcInfo();
-            const signerRA = await Verifier.fetSignerRA(svc.url, svc.model);
-            if (!signerRA?.signing_address) {
-                throw new Error('signing address does not exist');
+            let signerRA = {
+                signing_address: '',
+                nvidia_payload: '',
+                intel_quote: '',
+            };
+            if (vllmProxy) {
+                signerRA = await Verifier.fetSignerRA(svc.url, svc.model);
+                if (!signerRA?.signing_address) {
+                    throw new Error('signing address does not exist');
+                }
+            }
+            else {
+                let { quote, provider_signer, nvidia_payload } = await this.getQuote(providerAddress);
+                signerRA = {
+                    signing_address: provider_signer,
+                    nvidia_payload: nvidia_payload,
+                    intel_quote: quote,
+                };
             }
             signingKey = `${this.contract.getUserAddress()}_${providerAddress}`;
             await this.metadata.storeSigningKey(signingKey, signerRA.signing_address);
@@ -8790,6 +8686,132 @@ class Verifier extends ZGServingUserBrokerBase {
 }
 
 /**
+ * RequestProcessor is a subclass of ZGServingUserBroker.
+ * It needs to be initialized with createZGServingUserBroker
+ * before use.
+ */
+class RequestProcessor extends ZGServingUserBrokerBase {
+    automata;
+    constructor(contract, metadata, cache, ledger) {
+        super(contract, ledger, metadata, cache);
+        this.automata = new Automata();
+    }
+    async getServiceMetadata(providerAddress) {
+        const service = await this.getService(providerAddress);
+        return {
+            endpoint: `${service.url}/v1/proxy`,
+            model: service.model,
+        };
+    }
+    /*
+     * 1. To Ensure No Insufficient Balance Occurs.
+     *
+     * The provider settles accounts regularly. In addition, we will add a rule to the provider's settlement logic:
+     * if the actual balance of the customer's account is less than 5000, settlement will be triggered immediately.
+     * The actual balance is defined as the customer's inference account balance minus any unsettled amounts.
+     *
+     * This way, if the customer checks their account and sees a balance greater than 5000, even if the provider settles
+     * immediately, the deduction will leave about 5000, ensuring that no insufficient balance situation occurs.
+     *
+     * 2. To Avoid Frequent Transfers
+     *
+     * On the customer's side, if the balance falls below 5000, it should be topped up to 10000. This is to avoid frequent
+     * transfers.
+     *
+     * 3. To Avoid Having to Check the Balance on Every Customer Request
+     *
+     * Record expenditures in processResponse and maintain a total consumption amount. Every time the total expenditure
+     * reaches 1000, recheck the balance and perform a transfer if necessary.
+     *
+     * ps: The units for 5000 and 1000 can be (service.inputPricePerToken + service.outputPricePerToken).
+     */
+    async getRequestHeaders(providerAddress, content, vllmProxy) {
+        try {
+            await this.topUpAccountIfNeeded(providerAddress, content);
+            if (!vllmProxy) {
+                vllmProxy = false;
+            }
+            return await this.getHeader(providerAddress, content, BigInt(0), vllmProxy);
+        }
+        catch (error) {
+            throw error;
+        }
+    }
+    async acknowledgeProviderSigner(providerAddress, gasPrice) {
+        try {
+            try {
+                await this.contract.getAccount(providerAddress);
+            }
+            catch (error) {
+                if (!error.message.includes('AccountNotExists')) {
+                    throw error;
+                }
+                else {
+                    await this.ledger.transferFund(providerAddress, 'inference', BigInt(0), gasPrice);
+                }
+            }
+            let { quote, provider_signer, key, nvidia_payload } = await this.getQuote(providerAddress);
+            if (!quote || !provider_signer) {
+                throw new Error('Invalid quote');
+            }
+            if (!quote.startsWith('0x')) {
+                quote = '0x' + quote;
+            }
+            const rpc = process.env.RPC_ENDPOINT;
+            // bypass quote verification if testing on localhost
+            if (!rpc || !/localhost|127\.0\.0\.1/.test(rpc)) {
+                const isVerified = await this.automata.verifyQuote(quote);
+                console.log('Quote verification:', isVerified);
+                if (!isVerified) {
+                    throw new Error('Quote verification failed');
+                }
+                if (nvidia_payload) {
+                    const valid = await Verifier.verifyRA(nvidia_payload);
+                    console.log('nvidia payload verification:', valid);
+                    if (!valid) {
+                        throw new Error('nvidia payload verify failed');
+                    }
+                }
+            }
+            const account = await this.contract.getAccount(providerAddress);
+            if (account.providerPubKey[0] === key[0] &&
+                account.providerPubKey[1] === key[1]) {
+                console.log('Provider signer already acknowledged');
+                return;
+            }
+            await this.contract.acknowledgeProviderSigner(providerAddress, key);
+            const userAddress = this.contract.getUserAddress();
+            const cacheKey = `${userAddress}_${providerAddress}_ack`;
+            await this.cache.setItem(cacheKey, key, 1 * 60 * 1000, CacheValueTypeEnum.Other);
+        }
+        catch (error) {
+            throw error;
+        }
+    }
+}
+
+var VerifiabilityEnum;
+(function (VerifiabilityEnum) {
+    VerifiabilityEnum["OpML"] = "OpML";
+    VerifiabilityEnum["TeeML"] = "TeeML";
+    VerifiabilityEnum["ZKML"] = "ZKML";
+})(VerifiabilityEnum || (VerifiabilityEnum = {}));
+let ModelProcessor$1 = class ModelProcessor extends ZGServingUserBrokerBase {
+    async listService() {
+        try {
+            const services = await this.contract.listService();
+            return services;
+        }
+        catch (error) {
+            throw error;
+        }
+    }
+};
+function isVerifiability(value) {
+    return Object.values(VerifiabilityEnum).includes(value);
+}
+
+/**
  * ResponseProcessor is a subclass of ZGServingUserBroker.
  * It needs to be initialized with createZGServingUserBroker
  * before use.
@@ -8800,7 +8822,7 @@ class ResponseProcessor extends ZGServingUserBrokerBase {
         super(contract, ledger, metadata, cache);
         this.verifier = new Verifier(contract, ledger, metadata, cache);
     }
-    async processResponse(providerAddress, content, chatID, useProxy) {
+    async processResponse(providerAddress, content, chatID, vllmProxy) {
         try {
             const extractor = await this.getExtractor(providerAddress);
             const outputFee = await this.calculateOutputFees(extractor, content);
@@ -8812,18 +8834,18 @@ class ResponseProcessor extends ZGServingUserBrokerBase {
             if (!chatID) {
                 throw new Error('Chat ID does not exist');
             }
+            if (!vllmProxy) {
+                vllmProxy = false;
+            }
             let singerRAVerificationResult = await this.verifier.getSigningAddress(providerAddress);
             if (!singerRAVerificationResult.valid) {
                 singerRAVerificationResult =
-                    await this.verifier.getSigningAddress(providerAddress, true);
+                    await this.verifier.getSigningAddress(providerAddress, true, vllmProxy);
             }
             if (!singerRAVerificationResult.valid) {
                 throw new Error('Signing address is invalid');
             }
-            if (!useProxy) {
-                useProxy = false;
-            }
-            const ResponseSignature = await Verifier.fetSignatureByChatID(svc.url, chatID, svc.model, useProxy);
+            const ResponseSignature = await Verifier.fetSignatureByChatID(svc.url, chatID, svc.model, vllmProxy);
             return Verifier.verifySignature(ResponseSignature.text, ResponseSignature.signature, singerRAVerificationResult.signingAddress);
         }
         catch (error) {
