+++ src.ts/sdk/inference/broker/response.ts
@@ -28,7 +28,7 @@ export class ResponseProcessor extends ZGServingUserBrokerBase {
         providerAddress: string,
         content: string,
         chatID?: string,
-        useProxy?: boolean
+        vllmProxy?: boolean
     ): Promise<boolean | null> {
         try {
             const extractor = await this.getExtractor(providerAddress)
@@ -44,26 +44,31 @@ export class ResponseProcessor extends ZGServingUserBrokerBase {
                 throw new Error('Chat ID does not exist')
             }
 
+            if (!vllmProxy) {
+                vllmProxy = false
+            }
+
             let singerRAVerificationResult =
                 await this.verifier.getSigningAddress(providerAddress)
 
             if (!singerRAVerificationResult.valid) {
                 singerRAVerificationResult =
-                    await this.verifier.getSigningAddress(providerAddress, true)
+                    await this.verifier.getSigningAddress(
+                        providerAddress,
+                        true,
+                        vllmProxy
+                    )
             }
 
             if (!singerRAVerificationResult.valid) {
                 throw new Error('Signing address is invalid')
             }
 
-            if (!useProxy) {
-                useProxy = false
-            }
             const ResponseSignature = await Verifier.fetSignatureByChatID(
                 svc.url,
                 chatID,
                 svc.model,
-                useProxy
+                vllmProxy
             )
 
             return Verifier.verifySignature(
